api_mode/requirements_api.txt
# Flask Web Framework
Flask==3.0.0
flask-restx==1.3.0
Werkzeug==3.0.1

# API Documentation (Swagger UI)
Flask-CORS==4.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# Request/Response Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Configuration
PyYAML==6.0.1

# Database
# SQLite is built into Python, no need to install

# Utilities
python-dotenv==1.0.0

# HTTP Requests (for calling external APIs)
requests==2.31.0

# Date/Time utilities
python-dateutil==2.8.2

# Existing automation dependencies (from parent requirements.txt)
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Inherit parent requirements
# These are already in parent requirements.txt:
# - google-auth
# - google-auth-oauthlib
# - google-auth-httplib2
# - google-api-python-client
# - requests
# - schedule
